update code follow MVC model
separate to folder model, controller and route. Change folder name
