+++ Nguyen thanh dat/python/book CRUD/Controller/sort_algorithm.py
@@ -1,0 +1,7 @@
+
+def bubble_sort(list):
+    for i in range(0, len(list)):
+        for j in range(0, len(list)-i-1):
+            if list[j] > list[j+1]:
+                list[j], list[j+1] = list[j+1], list[j]
+    return list
