+++ Nguyen thanh dat/python/book CRUD/Model/__init__.py
@@ -1,0 +1,3 @@
+
+
+__all__ = ["Book"]
