+++ Nguyen thanh dat/python/book CRUD/templates/index.html
@@ -32,7 +32,7 @@
         {% endfor %}
     </table>
     <h2>Add Book</h2>
-    <form action="/receive_book" method="POST">
+    <form action="/add_book" method="POST">
         <label for="category">Category:</label>
         <input type="text" name="category" id="category"><br>
