--- Nguyen thanh dat/python/flask_mysql/book/__init__.py
@@ -1,3 +1,0 @@
-from .book import book_bp
-
-__all__ = ["book_bp"]
